move common functions to top level of project
